Move unix-specific code to own module in mullvad-paths
