+++ mullvad-paths/src/cache.rs
@@ -5,9 +5,9 @@ use std::{env, path::PathBuf};
 /// one if that variable is unset.
 pub fn cache_dir() -> Result<PathBuf> {
     #[cfg(target_os = "linux")]
-    let permissions = crate::Permissions::Any;
+    let permissions = crate::unix::Permissions::Any;
     #[cfg(target_os = "macos")]
-    let permissions = crate::Permissions::ReadExecOnly;
+    let permissions = crate::unix::Permissions::ReadExecOnly;
     #[cfg(target_os = "windows")]
     let permissions = true;
     crate::create_and_return(get_cache_dir()?, permissions)
@@ -28,7 +28,7 @@ pub fn get_default_cache_dir() -> Result<PathBuf> {
 
 #[cfg(windows)]
 pub fn get_default_cache_dir() -> Result<PathBuf> {
-    let dir = crate::get_allusersprofile_dir()?
+    let dir = crate::windows::get_allusersprofile_dir()?
         .join(crate::PRODUCT_NAME)
         .join("cache");
     Ok(dir)
